pyeth-proxy/pyproject.toml
[tool.poetry]
name = "pyeth-proxy"
version = "0.1.0"
description = ""
authors = ["Ayaz Abbas <lfcayaz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
requests = "^2.31.0"
python-logging-loki = "^0.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
